refactor: renamed context setter
